iOS: secondsFromGMT returns Long on arm64 — cast offset to Int

+++ shared/src/iosMain/kotlin/com/indiana/zwl/shared/ios/ForestApp.kt
@@ -429,5 +429,5 @@ class ForestApp(
     }
 
     private fun timeZoneOffsetMinutes(atMillis: Long): Int =
-        (NSTimeZone.systemTimeZone.secondsFromGMT) / 60
+        (NSTimeZone.systemTimeZone.secondsFromGMT / 60).toInt()
 }
